@@ -1,10 +1,9 @@
 package br.com.olivum.recorder.activity;
 
+import android.Manifest;
 import android.content.Intent;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
+import android.content.pm.PackageManager;
 import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -12,6 +11,13 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
+import android.widget.Toast;
+
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 
 import br.com.olivum.recorder.R;
 import br.com.olivum.recorder.audio.AudioListener;
@@ -119,6 +125,8 @@ public class AudioNoteViewActivity extends AppCompatActivity implements AudioLis
         });
 
         AudioNote.registerListener(this);
+
+        requestAudioPermissions();
     }
 
     public void updateButton() {
@@ -192,5 +200,55 @@ public class AudioNoteViewActivity extends AppCompatActivity implements AudioLis
         super.onBackPressed();
 
         stop();
+    }
+
+    private final int MY_PERMISSIONS_RECORD_AUDIO = 1;
+
+    private void requestAudioPermissions() {
+        // Check if permission is granted
+        if (ContextCompat.checkSelfPermission(this, Manifest.permission.RECORD_AUDIO) != PackageManager.PERMISSION_GRANTED) {
+
+            // Permission is not granted
+            if (ActivityCompat.shouldShowRequestPermissionRationale(this, Manifest.permission.RECORD_AUDIO)) {
+                Toast.makeText(this, "Please grant permissions to record audio", Toast.LENGTH_LONG).show();
+
+                // Request user permission
+                ActivityCompat.requestPermissions(this,
+                                                  new String[]{Manifest.permission.RECORD_AUDIO},
+                                                  MY_PERMISSIONS_RECORD_AUDIO);
+
+            }
+            else {
+                // Request user permission
+                ActivityCompat.requestPermissions(this,
+                                                  new String[]{Manifest.permission.RECORD_AUDIO},
+                                                  MY_PERMISSIONS_RECORD_AUDIO);
+            }
+        }
+        // permission granted
+        else if (ContextCompat.checkSelfPermission(this, Manifest.permission.RECORD_AUDIO) == PackageManager.PERMISSION_GRANTED) {
+            //
+        }
+    }
+
+    //Handling callback
+    @Override
+    public void onRequestPermissionsResult(int requestCode,
+                                           String permissions[],
+                                           int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+
+        switch (requestCode) {
+            case MY_PERMISSIONS_RECORD_AUDIO: {
+                if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                    // permission was granted
+                } else {
+                    // permission denied
+                    Toast.makeText(this, "Permissions Denied to record audio", Toast.LENGTH_LONG).show();
+                }
+
+                return;
+            }
+        }
     }
 }
